Improving Firefox compatibility by getting rid of outerHTML use
Firefox does not have the outerHTML property so, to make it possible
to get the generated content of DOM objects, I'm creating a temp
container, insertting generated objects in it and then returning its
innerHTML prop value.

A test still breaking in FF because of the order of attributes. Didn't find
a way to do it better yet.