Adding tests and parsing for more fields in atom.parseEntry
Now we also support published and summary fields. Tests were added
to authors name and mail address.

The heaviest change in this commit was in the way that we find tags in
an atom entry. Now, instead of looking for a single tag in different times,
now there's a loop with all tags and a switch decides what to do
depending on the tag name. I'm preparing the way for dealing
with extension fields.